parse hashes in urls
This adds two things the url parser used in the router:

1. Parsing of hashes from the full url path string
2. Stringifying a RootUrl object to include hashes

This means that routes defined with regexes can match against hashes,
where before they would only be tested against the part that was parsed.
This also means that when we successfully route, during the commit step,
we will be setting the hash in the url instead of clearing it every
time. All it took was making the code hash-aware.